chore: move prices cli into price
